Use separate messages for RTTY than CW
To enable starting with CR, add a | character that is translated to CR

Add config keys DKF1..DKF12, ALT_DK1..ALT_DK10, DKCQM, DKSPM, DKSPC

Default values are the CW message with | at the start and the trailing
newline changed to a space.

Note that in callinput.c, in SSB mode, ENTER on a dupe when NOB4 is not
set would send F7 message in CW.  This now sends the voice F7 message
instead since that was likely the intent.

New function send_standard_message(int) added which sends the passed
message number using the current trxmode (cw, voice, or digi).  Also
added is send_keyer_message(int) which works as above, but in voice
mode will still send CW.  This is used by the CW keyer.

LIST and MESSAGE will now show/edit the DIGI messages in DIGI mode.
In CW and voice modes, the CW messages will be shown/changed.